hw06-jpa/src/main/java/ru/otus/hw/repositories/JpaGenreRepository.java
package ru.otus.hw.repositories;

import jakarta.persistence.EntityManager;
import jakarta.persistence.PersistenceContext;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;
import ru.otus.hw.models.Genre;

import java.util.List;
import java.util.Set;

@Repository
@RequiredArgsConstructor
public class JpaGenreRepository implements GenreRepository {

    @PersistenceContext
    private final EntityManager em;

    @Override
    public List<Genre> findAll() {
        var queryString = "select g from Genre g";
        return em.createQuery(queryString, Genre.class).getResultList();
    }

    @Override
    public List<Genre> findAllByIds(Set<Long> ids) {
        var queryString = "select g from Genre g where g.id in (:ids)";
        return em.createQuery(queryString, Genre.class)
            .setParameter("ids", ids)
            .getResultList();
    }
}
